requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.4
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cnocr==2.3.1
cnstd==1.2.5.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
datasets==3.3.2
dill==0.3.8
dnspython==2.7.0
docker-pycreds==0.4.0
doclayout_yolo==0.0.3
email_validator==2.2.0
evaluate==0.4.3
fastapi==0.115.8
fastapi-cli==0.0.7
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.5
kiwisolver==1.4.8
latex2mathml==3.77.0
lightning-utilities==0.12.0
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
numpy==2.1.1
onnx==1.17.0
onnxruntime==1.20.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
optimum==1.24.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pix2text==1.1.2.3
platformdirs==4.3.6
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyMuPDF==1.25.3
pyparsing==3.2.1
pyreadline3==3.5.4
pyspellchecker==0.8.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
pytorch-lightning==2.5.0.post0
pytz==2025.1
PyYAML==6.0.2
rapidocr-onnxruntime==1.3.25
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
safetensors==0.5.2
scipy==1.15.2
seaborn==0.13.2
sentry-sdk==2.22.0
setproctitle==1.3.5
setuptools==75.8.0
shapely==2.0.7
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.45.3
stringzilla==3.11.3
sympy==1.13.1
thop==0.1.1.post2209072238
tokenizers==0.21.0
torch==2.6.0
torchmetrics==1.6.1
torchvision==0.21.0
tqdm==4.67.1
transformers==4.48.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.78
ultralytics-thop==2.0.14
Unidecode==1.3.8
urllib3==2.3.0
uvicorn==0.34.0
wandb==0.19.7
watchfiles==1.0.4
websockets==15.0
xxhash==3.5.0
yarl==1.18.3
